Rename KeymapProvider to ActionDispatcher
